leetcode/src/main/java/com/newzhxu/L206.java
package com.newzhxu;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

public class L206 {
    @Data
    @NoArgsConstructor
    @AllArgsConstructor
    public static class ListNode{
        int val;
        ListNode next;

    }
    // 通过迭代的方式反转链表
    public ListNode reverseList(ListNode head){
        ListNode prev = null;
        ListNode curr = head;
        while(curr != null){
            ListNode nextTemp = curr.next;
            curr.next = prev;
            prev = curr;
            curr = nextTemp;
        }
        return prev;
    }
    // 通过递归的方式反转链表
    public ListNode reverseList2(ListNode head){
        if(head == null || head.next == null){
            return head;
        }
        ListNode p = reverseList2(head.next);
        head.next.next = head;
        head.next = null;
        return p;
    }
}
